requirements.txt
google-genai
requests
pydantic
firebase-admin
python-dotenv
